Review and document module dub.dub.
Moves the module name determination code to dub.internal.utils and makes DependencyVersionResolver private, as that will probably change considerably in the future (no deprecation path for this, because it's unlikely to be used outside out DUB).

Deprecates Dub.list(String)ImportPaths, as these functions are fundamentally flawed, because they are taking just a subset of the possible fields in GeneratorSettings. listProjectData is the designated successor.